requirements.txt
fastapi==0.109.0
librosa==0.8.0
numpy==1.23.5
omegaconf==2.0.6
packaging==23.2
phonemizer==2.2.1
platformdirs==3.5.1
pooch==1.8.0
portalocker==2.8.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pylatexenc==2.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
resampy==0.4.2
rfc3986==1.5.0
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
segments==2.2.1
soundfile==0.12.1
tokenizers==0.15.1
torch
torchaudio
transformers==4.37.1
uvicorn==0.27.0
googletrans==3.1.0a0
openai==0.27.0
python-dotenv==1.0.0
easymms